serde_derive

Module de

Modules

enum_ 🔒

enum_adjacently 🔒

Deserialization for adjacently tagged enums:

enum_externally 🔒

Deserialization for externally tagged enums:

enum_internally 🔒

Deserialization for internally tagged enums:

enum_untagged 🔒

Deserialization for untagged enums:

identifier 🔒

Deserialization of struct field identifiers and enum variant identifiers by way of a Rust enum.

struct_ 🔒

tuple 🔒

unit 🔒

Structs

DeImplGenerics 🔒

DeTypeGenerics 🔒

FieldWithAliases 🔒

Parameters 🔒

Enums

BorrowedLifetimes 🔒

StructForm 🔒

TupleForm 🔒

Functions

borrowed_lifetimes 🔒

build_generics 🔒

de_type_generics_to_tokens 🔒

deserialize_body 🔒

deserialize_from 🔒

Generates Deserialize::deserialize body for a type with #[serde(from)] attribute

deserialize_in_place_body 🔒

deserialize_seq 🔒

deserialize_transparent 🔒

Generates Deserialize::deserialize body for a type with #[serde(transparent)] attribute

deserialize_try_from 🔒

Generates Deserialize::deserialize body for a type with #[serde(try_from)] attribute

effective_style 🔒

expand_derive_deserialize

expr_is_missing 🔒

expr_is_missing_seq 🔒

field_i 🔒

has_flatten 🔒

True if there is any field with a #[serde(flatten)] attribute, other than fields which are skipped.

needs_deserialize_bound 🔒

precondition 🔒

precondition_no_de_lifetime 🔒

precondition_sized 🔒

requires_default 🔒

unwrap_to_variant_closure 🔒

wrap_deserialize_field_with 🔒

wrap_deserialize_with 🔒

This function wraps the expression in #[serde(deserialize_with = "...")] in a trait to prevent it from accessing the internal Deserialize state.