Struct serde_derive::de::Parameters

struct Parameters {
    local: Ident,
    this: Path,
    generics: Generics,
    borrowed: BorrowedLifetimes,
    has_getter: bool,
}

Fields
local: Ident

Name of the type the derive is on.

this: Path

Path to the type the impl is for. Either a single Ident for local types or some::remote::Ident for remote types. Does not include generic parameters.

generics: Generics

Generics including any explicit and inferred bounds for the impl.

borrowed: BorrowedLifetimes

Lifetimes borrowed from the deserializer. These will become bounds on the 'de lifetime of the deserializer.

has_getter: bool

At least one field has a serde(getter) attribute, implying that the remote type has a private field.

Methods

impl Parameters

fn new(cont: &Container) -> Self

fn type_name(&self) -> &str

Type name to use in error messages and &'static str arguments to various Deserializer methods.