Module serde_derive::fragment

Structs

Expr
Interpolate a fragment in place of an expression. This involves surrounding Block fragments in curly braces.

Match
Interpolate a fragment as the value part of a match expression. This involves putting a comma after expressions and curly braces around blocks.

Stmts
Interpolate a fragment as the statements of a block.

Enums

Fragment