serde_derive::internals

Module attr

Re-exports

pub use crate::internals::case::RenameRule;

Structs

Attr 🔒

BoolAttr 🔒

BorrowAttribute 🔒

Container

Represents struct or enum attribute information.

Field

Represents field attribute information

RenameAllRules

Variant

Represents variant attribute information

VecAttr 🔒

Enums

Default

Represents the default to use for a field when deserializing.

Identifier

Whether this enum represents the fields of a struct or the variants of an enum.

TagType

Styles of representing an enum.

Functions

borrowable_lifetimes 🔒

collect_lifetimes 🔒

collect_lifetimes_from_tokens 🔒

decide_identifier 🔒

decide_tag 🔒

get_lit_str 🔒

get_lit_str2 🔒

get_multiple_renames 🔒

get_renames 🔒

get_ser_and_de 🔒

get_where_predicates 🔒

is_cow 🔒

is_implicitly_borrowed 🔒

is_implicitly_borrowed_reference 🔒

is_option 🔒

is_primitive_path 🔒

is_primitive_type 🔒

is_reference 🔒

is_slice_u8 🔒

is_str 🔒

parse_lit_into_expr_path 🔒

parse_lit_into_lifetimes 🔒

parse_lit_into_path 🔒

parse_lit_into_ty 🔒

parse_lit_into_where 🔒

unraw 🔒

Type Aliases

SerAndDe 🔒