serde_derive

Module de

Modules

enum_ ğŸ”’

enum_adjacently ğŸ”’

Deserialization for adjacently tagged enums:

enum_externally ğŸ”’

Deserialization for externally tagged enums:

enum_internally ğŸ”’

Deserialization for internally tagged enums:

enum_untagged ğŸ”’

Deserialization for untagged enums:

identifier ğŸ”’

Deserialization of struct field identifiers and enum variant identifiers by way of a Rust enum.

struct_ ğŸ”’

tuple ğŸ”’

unit ğŸ”’

Structs

DeImplGenerics ğŸ”’

DeTypeGenerics ğŸ”’

FieldWithAliases ğŸ”’

Parameters ğŸ”’

Enums

BorrowedLifetimes ğŸ”’

StructForm ğŸ”’

TupleForm ğŸ”’

Functions

borrowed_lifetimes ğŸ”’

build_generics ğŸ”’

de_type_generics_to_tokens ğŸ”’

deserialize_body ğŸ”’

deserialize_from ğŸ”’

Generates Deserialize::deserialize body for a type with #[serde(from)] attribute

deserialize_in_place_body ğŸ”’

deserialize_seq ğŸ”’

deserialize_transparent ğŸ”’

Generates Deserialize::deserialize body for a type with #[serde(transparent)] attribute

deserialize_try_from ğŸ”’

Generates Deserialize::deserialize body for a type with #[serde(try_from)] attribute

effective_style ğŸ”’

expand_derive_deserialize

expr_is_missing ğŸ”’

expr_is_missing_seq ğŸ”’

field_i ğŸ”’

has_flatten ğŸ”’

True if there is any field with a #[serde(flatten)] attribute, other than fields which are skipped.

needs_deserialize_bound ğŸ”’

precondition ğŸ”’

precondition_no_de_lifetime ğŸ”’

precondition_sized ğŸ”’

requires_default ğŸ”’

unwrap_to_variant_closure ğŸ”’

wrap_deserialize_field_with ğŸ”’

wrap_deserialize_with ğŸ”’

This function wraps the expression in #[serde(deserialize_with = "...")] in a trait to prevent it from accessing the internal Deserialize state.