Module serde_derive::de

Structs

DeImplGenerics ğŸ”’

DeTypeGenerics ğŸ”’

Parameters ğŸ”’

Enums

BorrowedLifetimes ğŸ”’

StructForm ğŸ”’

TupleForm ğŸ”’

Functions

borrowed_lifetimes ğŸ”’

build_generics ğŸ”’

de_type_generics_to_tokens ğŸ”’

deserialize_adjacently_tagged_enum ğŸ”’

deserialize_body ğŸ”’

deserialize_custom_identifier ğŸ”’

deserialize_enum ğŸ”’

deserialize_externally_tagged_enum ğŸ”’

deserialize_externally_tagged_newtype_variant ğŸ”’

deserialize_externally_tagged_variant ğŸ”’

deserialize_field_identifier ğŸ”’

Generates enum and its Deserialize implementation that represents each non-skipped field of the struct

deserialize_from ğŸ”’

deserialize_generated_identifier ğŸ”’

deserialize_homogeneous_enum ğŸ”’

deserialize_identifier ğŸ”’

deserialize_in_place_body ğŸ”’

deserialize_internally_tagged_enum ğŸ”’

deserialize_internally_tagged_variant ğŸ”’

deserialize_map ğŸ”’

deserialize_newtype_struct ğŸ”’

deserialize_seq ğŸ”’

deserialize_struct ğŸ”’

deserialize_transparent ğŸ”’

deserialize_try_from ğŸ”’

deserialize_tuple ğŸ”’

deserialize_unit_struct ğŸ”’

deserialize_untagged_enum ğŸ”’

deserialize_untagged_enum_after ğŸ”’

deserialize_untagged_newtype_variant ğŸ”’

deserialize_untagged_variant ğŸ”’

effective_style ğŸ”’

expand_derive_deserialize

expr_is_missing ğŸ”’

expr_is_missing_seq ğŸ”’

field_i ğŸ”’

needs_deserialize_bound ğŸ”’

precondition ğŸ”’

precondition_no_de_lifetime ğŸ”’

precondition_sized ğŸ”’

prepare_enum_variant_enum ğŸ”’

requires_default ğŸ”’

split_with_de_lifetime ğŸ”’

unwrap_to_variant_closure ğŸ”’

wrap_deserialize_field_with ğŸ”’

wrap_deserialize_variant_with ğŸ”’

wrap_deserialize_with ğŸ”’

This function wraps the expression in #[serde(deserialize_with = "...")] in a trait to prevent it from accessing the internal Deserialize state.