Module serde_derive::de

Structs

DeImplGenerics 🔒

DeTypeGenerics 🔒

Parameters 🔒

Enums

BorrowedLifetimes 🔒

StructForm 🔒

TupleForm 🔒

Functions

borrowed_lifetimes 🔒

build_generics 🔒

de_type_generics_to_tokens 🔒

deserialize_adjacently_tagged_enum 🔒

deserialize_body 🔒

deserialize_custom_identifier 🔒

deserialize_enum 🔒

deserialize_externally_tagged_enum 🔒

deserialize_externally_tagged_newtype_variant 🔒

deserialize_externally_tagged_variant 🔒

deserialize_field_identifier 🔒

Generates enum and its Deserialize implementation that represents each non-skipped field of the struct

deserialize_from 🔒

deserialize_generated_identifier 🔒

deserialize_homogeneous_enum 🔒

deserialize_identifier 🔒

deserialize_in_place_body 🔒

deserialize_internally_tagged_enum 🔒

deserialize_internally_tagged_variant 🔒

deserialize_map 🔒

deserialize_newtype_struct 🔒

deserialize_seq 🔒

deserialize_struct 🔒

deserialize_transparent 🔒

deserialize_try_from 🔒

deserialize_tuple 🔒

deserialize_unit_struct 🔒

deserialize_untagged_enum 🔒

deserialize_untagged_enum_after 🔒

deserialize_untagged_newtype_variant 🔒

deserialize_untagged_variant 🔒

effective_style 🔒

expand_derive_deserialize

expr_is_missing 🔒

expr_is_missing_seq 🔒

field_i 🔒

needs_deserialize_bound 🔒

precondition 🔒

precondition_no_de_lifetime 🔒

precondition_sized 🔒

prepare_enum_variant_enum 🔒

requires_default 🔒

split_with_de_lifetime 🔒

unwrap_to_variant_closure 🔒

wrap_deserialize_field_with 🔒

wrap_deserialize_variant_with 🔒

wrap_deserialize_with 🔒

This function wraps the expression in #[serde(deserialize_with = "...")] in a trait to prevent it from accessing the internal Deserialize state.