Module serde_derive::internals::attr

Re-exports

pub use crate::internals::case::RenameRule;

Structs

Attr ğŸ”’

BoolAttr ğŸ”’

BorrowAttribute ğŸ”’

Container

Represents struct or enum attribute information.

Field

Represents field attribute information

Name

RenameAllRules

Variant

Represents variant attribute information

VecAttr ğŸ”’

Enums

Default

Represents the default to use for a field when deserializing.

Identifier

Whether this enum represents the fields of a struct or the variants of an enum.

TagType

Styles of representing an enum.

Functions

borrowable_lifetimes ğŸ”’

collect_lifetimes ğŸ”’

collect_lifetimes_from_tokens ğŸ”’

decide_identifier ğŸ”’

decide_tag ğŸ”’

get_lit_str ğŸ”’

get_lit_str2 ğŸ”’

get_multiple_renames ğŸ”’

get_renames ğŸ”’

get_ser_and_de ğŸ”’

get_where_predicates ğŸ”’

is_cow ğŸ”’

is_implicitly_borrowed ğŸ”’

is_implicitly_borrowed_reference ğŸ”’

is_option ğŸ”’

is_primitive_path ğŸ”’

is_primitive_type ğŸ”’

is_reference ğŸ”’

is_slice_u8 ğŸ”’

is_str ğŸ”’

parse_lit_into_expr_path ğŸ”’

parse_lit_into_lifetimes ğŸ”’

parse_lit_into_path ğŸ”’

parse_lit_into_ty ğŸ”’

parse_lit_into_where ğŸ”’

unraw ğŸ”’

Type Aliases

SerAndDe ğŸ”’